Moved as much as code as possible out from the Harmony patches. No functional change.
